+++ src/components/ProtectedRoute/ProtectedRoute.jsx
@@ -4,8 +4,8 @@ import Preloader from '../Preloader/Preloader';
 
 const ProtectedRoute = ({ onlyUnAuth = false, children, ...rest }) => {
 
-  const user  = useSelector(store => store.getUser.user);
-  const userAuthCheck = useSelector(store => store.getUser.userAuthCheck)
+  const user  = useSelector(store => store.user.user);
+  const userAuthCheck = useSelector(store => store.user.userAuthCheck)
   const location = useLocation();
 
   if (!userAuthCheck) {
